[+] move repeated `GetTest---()` to shared testutils package
