+++ cmd/csv_receiver/csv_receiver_test.go
@@ -5,23 +5,9 @@ import (
 	"os"
 	"testing"
 
-	"github.com/destrex271/pgwatch3_rpc_server/sinks/pb"
+	testutils "github.com/destrex271/pgwatch3_rpc_server/sinks/test_utils"
 	"github.com/stretchr/testify/assert"
-	"google.golang.org/protobuf/types/known/structpb"
 )
-
-func GetTestMeasurementEnvelope() *pb.MeasurementEnvelope {
-	st, err := structpb.NewStruct(map[string]any{"key": "val"})
-	if err != nil {
-		panic(err)
-	}
-	measurements := []*structpb.Struct{st}
-	return &pb.MeasurementEnvelope{
-		DBName:           "test",
-		MetricName:       "testMetric",
-		Data:             measurements,
-	}
-}
 
 func TestUpdateMeasurements(t *testing.T) {
 	fullPath, err := os.Getwd()
@@ -31,7 +17,7 @@ func TestUpdateMeasurements(t *testing.T) {
 	recv := NewCSVReceiver(fullPath)
 
 	// Call Update Measurements with dummy data
-	msg := GetTestMeasurementEnvelope()
+	msg := testutils.GetTestMeasurementEnvelope()
 	_, err = recv.UpdateMeasurements(context.Background(), msg)
 	assert.NoError(t, err)
 	defer func() { _ = os.RemoveAll(fullPath + "/" + msg.DBName) }()
