+++ cmd/parquet_receiver/parquet_receiver_test.go
@@ -5,23 +5,9 @@ import (
 	"os"
 	"testing"
 
-	"github.com/destrex271/pgwatch3_rpc_server/sinks/pb"
+	testutils "github.com/destrex271/pgwatch3_rpc_server/sinks/test_utils"
 	"github.com/stretchr/testify/assert"
-	"google.golang.org/protobuf/types/known/structpb"
 )
-
-func GetTestMeasurementEnvelope() *pb.MeasurementEnvelope {
-	st, err := structpb.NewStruct(map[string]any{"key": "val"})
-	if err != nil {
-		panic(err)
-	}
-	measurements := []*structpb.Struct{st}
-	return &pb.MeasurementEnvelope{
-		DBName:           "test",
-		MetricName:       "testMetric",
-		Data:             measurements,
-	}
-}
 
 func TestUpdateMeasurements(t *testing.T) {
 	fullPath, err := os.Getwd()
@@ -30,7 +16,7 @@ func TestUpdateMeasurements(t *testing.T) {
 	}
 	defer func () { _ = os.RemoveAll(fullPath + "/parquet_readings") }()
 
-	msg := GetTestMeasurementEnvelope()
+	msg := testutils.GetTestMeasurementEnvelope()
 	recv := NewParquetReceiver(fullPath)
 	_, err = os.Stat(fullPath + "/parquet_readings")
 	assert.False(t, os.IsNotExist(err), "Measurements Directory does not exist")
